Changes involve the `PrinterDataServiceTest.cs` and the `PrinterSettings.razor` file.
Several properties were removed from the class and instead added to the `Printer` object. A new method `RemovePrinter` was added to remove a printer from the `Printers` list. A new method `GenerateUniquePrinterId` was added to generate a unique ID for each printer. The `CreatePrinterProfile` method was updated to use the properties of the `Printer` object instead of the class itself. Two new methods `SavePrinterData` and `LoadData` were added to save and load printer data to and from a JSON file.